Move all location-importing logic into Importer
The import_locations tasks is now just a thin wrapper around
`ecoroofs.locations.Importer`.

This also changes the importer so that overwriting just deletes records
from the tables defined in the `locations` app instead of recreating or
resetting the database. If the schema is changed significantly,
re-creating the database or running migrations will have to be done
manually.

ECO-22