Add BaseModel.name CharField
For consistency and to reduce some duplication, add a name CharField to
BaseModel. By default, order by this field and use it as the default
__str__() value.

In addition, remove the Page.title CharField and use the name field
inherited from BaseModel instead. This also allows the overridden
Page.slug field to be removed.