Replace non-ASCII chars in locations.csv
Replace things like fancy quotes with ASCII quotes. This uses the
unidecode tool. When we get new versions of the locations file from the
client, it can be fixed up like this:

    unidecode locations.csv > new-locations.csv
    mv new-locations.csv locations.csv

This is necessary so Python's CSV reader doesn't blow up (although,
strangely, there's only a problem with this when running the locations
importer in Docker, even though it's running the same version of
Python as locally).