Fix dev database user
This is confusing because we're trying to support old-style development
on the local host and Docker at the same time.

The default for old-style development is to use the local user's
username. The default for Docker-style development is to use the
project's package name.

It would probably be best to use the same--the package name--for both,
but that requires an extra step somewhere in `make init` to create that
user as a superuser.

Hm...