Revamp ecoroofs-postgres Docker image
Base on CentOS 6.7 image to match production more closely. This also
fixes an issue with trying to install PostGIS 2.1 into the official
postgres image (which is based on Debian Jessie).

The core of this could be abstracted out and reused in other projects
(assuming we want our Docker images to continue to match production for
the time being).